+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - workflow: configurable number of datasamples to use for Mnist (#264)
 
+### Changed
+
+- feat: rename node to organization (#282)
+
 ## [0.21.0] - 2022-06-07
 
 ### Added